refactor: extract embedded tests + delete dynamic.rs
Move #[cfg(test)] blocks from client.rs, instruction_arg.rs, and
lib.rs to lang/tests/ as integration tests. Delete dynamic.rs
(Pod transition complete) and update prelude to re-export PodString
and PodVec directly. CPI and pda_precompute tests stay in-file due
to pub(crate) dependencies.

-721 lines from audit surface (14,285 → 13,564).
