+++ lang/src/lib.rs
@@ -43,8 +43,11 @@ extern crate self as quasar_lang;
 /// Breaking changes to this module are not considered semver violations.
 #[doc(hidden)]
 pub mod __internal {
-    pub use solana_account_view::{
-        AccountView, RuntimeAccount, MAX_PERMITTED_DATA_INCREASE, NOT_BORROWED,
+    pub use {
+        crate::entrypoint::parse_one_account,
+        solana_account_view::{
+            AccountView, RuntimeAccount, MAX_PERMITTED_DATA_INCREASE, NOT_BORROWED,
+        },
     };
 
     // Header layout (little-endian u32):
@@ -97,7 +100,6 @@ pub mod context;
 pub mod cpi;
 /// Marker types for dynamic fields (`String<N>`, `Vec<T, N>`) and codec
 /// helpers.
-pub mod dynamic;
 /// Runtime helpers for Pod-based dynamic account fields (`PodString`, `PodVec`).
 pub mod pod_dynamic;
 /// Program entrypoint macros (`dispatch!`, `no_alloc!`, `panic_handler!`).
@@ -248,52 +250,3 @@ pub fn abort_program() -> ! {
     panic!("program aborted");
 }
 
-#[cfg(test)]
-mod tests {
-    use {super::*, solana_address::Address};
-
-    #[test]
-    fn keys_eq_identical() {
-        let a = Address::new_from_array([0xAB; 32]);
-        assert!(keys_eq(&a, &a));
-    }
-
-    #[test]
-    fn keys_eq_first_word_mismatch() {
-        let a = Address::new_from_array([0xFF; 32]);
-        let mut b_bytes = [0xFF; 32];
-        b_bytes[0] = 0x00;
-        let b = Address::new_from_array(b_bytes);
-        assert!(!keys_eq(&a, &b));
-    }
-
-    #[test]
-    fn keys_eq_last_word_mismatch() {
-        let a = Address::new_from_array([0xFF; 32]);
-        let mut b_bytes = [0xFF; 32];
-        b_bytes[31] = 0x00;
-        let b = Address::new_from_array(b_bytes);
-        assert!(!keys_eq(&a, &b));
-    }
-
-    #[test]
-    fn keys_eq_all_zero() {
-        let a = Address::new_from_array([0; 32]);
-        let b = Address::new_from_array([0; 32]);
-        assert!(keys_eq(&a, &b));
-    }
-
-    #[test]
-    fn is_system_program_zero() {
-        let addr = Address::new_from_array([0; 32]);
-        assert!(is_system_program(&addr));
-    }
-
-    #[test]
-    fn is_system_program_nonzero() {
-        let mut bytes = [0u8; 32];
-        bytes[16] = 1;
-        let addr = Address::new_from_array(bytes);
-        assert!(!is_system_program(&addr));
-    }
-}
